+++ src/components/organisms/DropDownIncidentsTypes/dropdownIncidentsTypese.scss
