+++ src/components/organisms/commingIssue/commingIssue.jsx
@@ -1,15 +1,16 @@
-import React from "react"
-import Title from "../../atoms/title/title"
-import IssueInformation from "../../molecules/issueInformation/issueInformation"
-import Card from "../../templates/card"
+import React from 'react';
+import Title from '../../atoms/title/title'
+import IssueInformation from '../../molecules/issueInformation/issueInformation';
 
 const CommingIssue = () => {
-  return (
-    <Card className="commingIssue">
-      <Title cssClass="card-title">2 évènements à venir</Title>
-      <IssueInformation />
-    </Card>
-  );
-};
+	return (
+		<div className="commingIssue">
+			<Title cssClass="card-title">
+				2 évènements à venir
+			</Title>
+			<IssueInformation />
+		</div>
+	)
+}
 
-export default CommingIssue;
+export default CommingIssue
