+++ src/components/organisms/commingIssue/commingIssue.scss
@@ -1,5 +1,7 @@
 .commingIssue {
-	.title-component {
-		margin-bottom: $large-space;
-	}
+	padding: $xx-large-space;
+	border: 1px solid $gray-lighter;
+	border-radius: $card-radius;
+	box-shadow: $outer-shadow;
+	background-color: $primary-utility-color;
 }
